requirements/base.txt
python-slugify==8.0.1  # https://github.com/un33k/python-slugify
Pillow==10.2.0  # https://github.com/python-pillow/Pillow
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.6.0  # https://github.com/evansd/whitenoise
redis==5.0.1  # https://github.com/redis/redis-py
hiredis==2.3.2  # https://github.com/redis/hiredis-py
celery==5.3.6
uvicorn[standard]==0.25.0  # https://github.com/encode/uvicorn

# Telegram
# ------------------------------------------------------------------------------
python-telegram-bot[job-queue]==20.4
web3==6.9.0
requests==2.31.0
loguru==0.5.3
asyncio==3.4.3

# Enviroment Manager
# ------------------------------------------------------------------------------
python-decouple==3.8  # https://pypi.org/project/python-decouple/

# Coinbase
# ------------------------------------------------------------------------------
# https://coinbase.github.io/coinbase-advanced-py/
# https://docs.cloud.coinbase.com/advanced-trade-api/docs/auth#cloud-api-keys
coinbase-advanced-py==1.2.0


# async functions
asgiref==3.8.1
